requirements.txt
# Site Mapper Dependencies
crawl4ai
google-generativeai
python-dotenv
beautifulsoup4
lxml
tldextract
